Bring some of Wasmtime's documentation more up-to-date (#6994)
* Remove tutorial/wasm-writing documentation

This commit removes the tutorial and "Writing WebAssembly" documentation
sections from Wasmtime's documentation. These sections are quite dated
at this point (they still recommend `cargo wasi`!) and haven't been
updated much since their inception, especially in the arena of
components. Today it seems best to leave this sort of documentation to
[other resources] which are more tailored towards documentation of
writing wasm.

[other resources]: https://component-model.bytecodealliance.org/

* Remove markdown example from docs

Like the previous commit this is quite dated and recommends
effectively-deprecated tooling and doesn't take into account components.

* Update our intro docs a bit

* Link security blog post from docs

* Fold docs of wasm proposals into tier docs

This was already a bit duplicated so consolidate into one location.
Additionally add some proposals that weren't previously documented and
move some around based on their implementation status.

* Document unsupported features

In an effort to head off questions about platform support I figure it
might be a good idea to start documenting what's not supported at this
time. This is intended to mirror the current state, not future, of
Wasmtime. In other words this should answer the question of "Does
Wasmtime support X?" as opposed to "Does Wasmtime want to support X?"
since we want to eventually support all of these features in the limit.

* Fold WASI docs into tier docs

Similar to the previous commit but for WASI proposals.