Add license to c-api macros
Running cargo license on the c-api crate reports that this crate is
unlicensed.

We use cargo-license to automate our license tracking for this library
in our C++ application and it'd be nice to just have a license in here.

I used the standard license that wasmtime uses, but happy to use
another.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>