wasi-cap-std-sync: use std::io::Std{in,out,err} to impl {read,write}_vectored (#6825)
* add new wasi-test printing unicode on stdout

there's not yet a way to show this test is actually doing what we
intend, I'm using CI to see how this runs on all platforms

Reported in #6824

prtest:full

* cap-std-sync: use std::io::Std{in,out,err}'s {read,write}_vectored directly

instead of using AsFilelike to view these as a File.

This means we also have to use std's mutexes to access these descriptors,
which also seems like a good idea.

We believe will fix unicode output on windows issues:
https://github.com/bytecodealliance/wasmtime/issues/6824