riscv64: Emit islands before `br_table` sequence
This fixes a slightly subtle issue with our island emission in BrTable.

We used to emit islands right before the jump table targets. This
causes issues because if the island is actually emitted, we have
the potential to jump right into the middle of it. This happens
because we have calculated a fixed offset from the `auipc` instruction
assuming no island is emitted.

This commit changes the island to be emitted right at the start of the
br_table sequence so that this cannot happen.