wasi: Fix a few issues around stdin
This commit is intended to address #6986 and some other issues related
to stdin and reading it, notably:

* Previously once EOF was reached the `closed` flag was mistakenly not
  set.
* Previously data would be infinitely buffered regardless of how fast
  the guest program would consume it.
* Previously stdin would be immediately ready by Wasmtime regardless of
  whether the guest wanted to read stdin or not.
* The host-side preview1-to-preview2 adapter didn't perform a blocking
  read meaning that it never blocked.

These issues are addressed by refactoring the code in question.
Note that this is similar to the logic of `AsyncReadStream` somewhat but
that type is not appropriate in this context due to the singleton nature
of stdin meaning that the per-stream helper task and per-stream buffer
of `AsyncReadStream` are not appropriate.

Closees #6986