Fix character boundary issues in preview1 host adapter (#7011)
* Fix character boundary issues in preview1 host adapter

This fixes two separate issues in the preview1-to-preview2 host-side
adapter in the `wasmtime-wasi` crate. Both instances were copying a
truncated string to the guest program but the truncation was happening
in the middle of a unicode character so truncation now happens on bytes
instead of a string.

* Update `write_bytes` to take `&[u8]`

Try to weed out any accidental string-related issues for the future.