Refactor how `subscribe` works in WASI
This commit refactors how the `Pollable` resource is created in WASI.
Previously this was created manually via two variants of a `Pollable`
enum but this was somewhat error prone for a few reasons:

* The more common representation, `TableEntry`, had boilerplate
  associated with it that had dynamic downcasts and such. This is now
  all hidden behind a more typed API.

* Callers had to remember to use `push_child_resource` which is easy to
  forget.

* The previous signature of the readiness check returned a `Result<()>`
  which made it accidentally easy for errors to be propagated into traps
  rather than being returned as a "last operation failed" error.

This commit replaces the previous API with a single `subscribe` function
which takes a `Resource<T>` and returns a `Resource<Pollable>`. The `T`
type must implement a new trait, `Subscribe`, which indicates the
asynchronous readiness check. This readiness check additionally returns
`()` so it's no longer possible to accidentally return errors (or trap).

This namely required refactoring the `HostOutputStream` trait and
implementations. The trait itself now has a `check_write` method
corresponding to to the same WASI function. The old `write_ready`
function is now a convenience helper combo between `ready` and
`check_write`. Concrete implementations are refactored to store errors
discovered during `ready` to get communicated through `check-write`
later on.