zkasm: delete a boatload of RISC-V specific code (#10)
* zkasm: delete a boatload of RISC-V specific code

A downside of doing it the "copy riscv" way is that we have a ton of
types and stuff that are specific to how RISC-V architecture works. A
prime example of that is `AluRRImm12` type – RISC-V instruction encoding
does not support immediates that are larger than 12 bits in most cases,
so the backend needs to keep that in mind.

In principle if we did it from scratch instead we'd avoid having all of
the optimization-oriented special cases and rules, and would be able to
specify just the rules how to convert operations between two registers
and such.

* Implement LoadConst32/64