riscv64: Move `sextend` optimizations into `sext` (#7156)
This commit moves the optimizations from `sextend` for various shapes of
instructions into the `sext` helper. This enables these optimizations to
kick in when generating other instructions which require sign-extensions
such as `icmp` and `brif`.

This additionally fixes an issue introduced in #7121 where
uextend-of-sextend wasn't translated correctly and the outer `uextend`
was mistakenly omitted.